requirements.txt
fastapi
uvicorn
motor
apscheduler
python-dotenv
requests
